Relocate 5 verifier note/diagnostics helpers to verifier_orchestration
Phase 2 batch 3 for Issue #682. Moves 5 medium pure-fn verifier
helpers (~180 LOC of bodies) out of turn.rs into
verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  System notes:
    - task_contract_verifier_repair_note (48 LOC)
    - verifier_repair_diagnostic_pending_note (37 LOC)
    - task_contract_verifier_targeted_edit_required_note (45 LOC)

  Diagnostic formatters:
    - verifier_repair_assessment_diagnostics (17 LOC)
    - verifier_repair_context_diagnostics (31 LOC)

All 5 are promoted to pub(super) (file-private before). The note
helpers all receive a `&RepairJob` context and emit prompt-injection-
sanitized system notes per Security Invariants (mask_secrets +
json-serialize escaping); behavior unchanged.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::VerifierRepairAssessment (parent loop_run module)
    - super::repair_job::verifier_repair_effective_target_hint
    - crate::safety::path_guard::resolve_user_path
    - crate::session::feedback::mask_secrets
    (BehaviorContractProjection and RepairJob were already imported).

  * turn.rs: extended use super::verifier_orchestration::{...} with
    the 5 newly-relocated production fns. verifier_repair_assessment_-
    diagnostics is only referenced by the moved
    verifier_repair_context_diagnostics, so it stays pub(super) but
    is not re-imported into turn.rs.

Metrics: turn.rs 34,742 -> 34,561 (-181); verifier_orchestration.rs
296 -> 484 (+188). Cumulative Issue #682 progress: turn.rs
34,958 -> 34,561 (-397 across PRs #714 / #715 / #716 / this PR).
Acceptance gap: 34,561 -> 32,000 = ~2,561 LOC remaining.
